requirements.txt
fastapi
uvicorn
jinja2
httpx
pytest
python-dotenv
python-multipart
mangum
python-socketio
sqlalchemy
pytest-cov
requests